+++ priv/test_repo/migrations/20260730000001_upgrade_system_flags_v03.exs
@@ -1,0 +1,6 @@
+defmodule PhoenixFlags.TestRepo.Migrations.UpgradeSystemFlagsV03 do
+  use Ecto.Migration
+
+  def up, do: PhoenixFlags.Migration.up(version: 3)
+  def down, do: PhoenixFlags.Migration.down(version: 3)
+end